==> 4a-rename.py <==
import os

os.rename('output_got.csv', 'output_got-santos.csv')
os.rename('tweets_pp_translation.csv', 'tweets_pp_translation-santos.csv')
os.rename('stats.txt', 'stats-santos.txt')
os.rename('tweets_pp.csv', 'tweets_pp-santos.csv')
