==> 4b-rename.py <==
import os

os.rename('output_got.csv', 'output_got-uribe.csv')
os.rename('tweets_pp_translation.csv', 'tweets_pp_translation-uribe.csv')
os.rename('stats.txt', 'stats-uribe.txt')
os.rename('tweets_pp.csv', 'tweets_pp-uribe.csv')
